+++ jpsi_pi0/scripts/full_sim.py
@@ -28,20 +28,22 @@ def execute_root(step, log, prev_out, prev_step, args=None):
     proc = subprocess.Popen("root -b -q %s" % args,shell=True, stdout=log, stderr=log)
     proc.wait()
     #test_exec(step,log,prev_step,prev_out,args)
-    #os.unlink("%s_complete.root"%prev_step)
     
 def execute_step(step, prev_out, prev_step):
     my_utils.dbg_msg( "====== executing step %s with uniq id %d " % (step,my_utils.uniq_id) )
     (out_file,log_file,dummy) = my_utils.file_names(step)
-    if (step != "sim"):
-        root_args =  "%s/%s_complete.C" % (my_utils.macro_dir,step)
-        with open(log_file,"w") as log:
-            execute_root(step, log, prev_out, prev_step, root_args)
+    if not os.path.exists(out_file):
+        if (step != "sim"):
+            root_args =  "%s/%s_complete.C" % (my_utils.macro_dir,step)
+            with open(log_file,"w") as log:
+                execute_root(step, log, prev_out, prev_step, root_args)
+        else:
+            root_args = "\'%s/%s_complete.C(\"%s\")\'" % (my_utils.macro_dir, step, prev_out)
+            with open(log_file,"w") as log:
+                execute_root(step, log, prev_out, prev_step, root_args)
+        my_utils.move_file("%s_complete.root" % step, out_file)
     else:
-        root_args = "\'%s/%s_complete.C(\"%s\")\'" % (my_utils.macro_dir, step, prev_out)
-        with open(log_file,"w") as log:
-            execute_root(step, log, prev_out, prev_step, root_args)
-    my_utils.move_file("%s_complete.root" % step, out_file)
+        my_utils.dbg_msg("%s exists. will skip this step (%s) " % (out_file,step))
     return (step,out_file)
 
 def run():
